requirements.txt
streamlit
polars
google-generativeai
PyMuPDF
tqdm
numpy